fix(ci): resolve all clippy warnings and formatting errors for zero-warning CI
- Fix manual_strip in header_anchor.rs (use strip_prefix)
- Fix unnecessary_to_owned in storage_proof.rs
- Fix derivable_impls for VerificationMode::Default
- Fix unused_enumerate_index and unnecessary_unwrap in rpc.rs
- Remove unused serde imports in defi-protocol demo
- Apply cargo clippy --fix for needless borrows across p2p and cli
- Apply cargo fmt for consistent formatting
- Add #[allow] for pre-existing issues that need refactoring:
  - collapsible_match in p2p client and monitor
  - dead_code, type_complexity, too_many_arguments in daemon
- Fix test_empty_tree_root_matches_sha256_empty -> keccak256

+++ sods-cli/src/api/health.rs
@@ -54,8 +54,8 @@ impl FromRef<Arc<AgentMetrics>> for MonitoringState {
 
 pub async fn health_check(State(state): State<MonitoringState>) -> Json<HealthResponse> {
     let (status, erc8004_status, health_metrics) = if let Some(ref metrics) = state.metrics {
-        let total_requests = metrics.validation_requests_received_total.get() as f64;
-        let successful_responses = metrics.validation_responses_submitted_total.get() as f64;
+        let total_requests = metrics.validation_requests_received_total.get();
+        let successful_responses = metrics.validation_responses_submitted_total.get();
         let success_rate = if total_requests > 0.0 {
             (successful_responses / total_requests) * 100.0
         } else {
